fix: moved `ParametrizationForm` input logic to `Parametrization` page
